[ConvertToLLVM] Don't choke on alloc of memref of index (#14002)
Prior to this patch the compiler would crash on alloc of memrefs of
index on shared memory, because we would try to get the bitwidth of the
index type as if it was an integer or a floating point type.

The bitwidth of the index type is actually carried by the datalayout of
the module.

Update the code to query the datalayout to get the bitwidth.

Note: This particular pattern is called as part of convert-to-nvvm. In
that pass we actually lower the index type to an integer type, but we do
it after having lowered the `alloc`s from shared memory, i.e., where the
code is choking.