Harden the MatmulTensorCore strategy (#13971)
This revision improves the behavior of the matmul tensor core strategy
for aligned and unaligned cases as follows:
- provide better behavior around using the maximum threads for copies by
avoiding rank-reducing patterns on vectors that trip up late
vector-to-mma conversions.
- plug the barrier removal pass added in #13886 : this removes an
unnecessary barrier between copies to shared memory and allows a form
that the mma.sync pipeline accepts without choking.
- force the synchronization of the sequential loop before hoisting
redundant transfers to workaround issue #13964. This addresses
synchronization issues reported in #13451 and #13811.
- various NFC cleanups and error message improvements.
- addition of failsafe configuration alterations to avoid hitting corner
cases where we would fail to tile and generate sequential code.
- turning on mma.sync by default in the unaligned and partially aligned
cases.

The changes have been tested against other implementations for numerical
accuracy and against races in a variety of unaligned/partially aligned
and aligned
[cases](https://github.com/iree-org/iree-samples/blob/f4ea2d5e5d0c0514a3d88a1926ffa6fc2d721e9f/transform_dialect/python/matmul_test.py)
with various compiler configurations.