[iree-dialects][test] Remove dead code from the transfrom script (#14127)
Dead code in the transform script can currently trigger a bug in the
interpreter. The issue is the transforms are applied to the transform
script as well.
For this particular test, we have a dead matcher that gets removed by
dce and this crashes the interpreter because it tries to access a
transform op that doesn't exist anymore (use-after-free).

While the interpreter is being fixed (upstream), remove the dead code to
avoid the crash.