Reserve extra scratch registers for >4byte remappings. (#14332)
Currently, if a branch remapping includes 64bit ints, it depends on the
topology of the program as to whether enough scratch space happens to be
available. In a case tested, it was not. This is because remappings are
done 32bits at a time and presume that scratch space has been reserved
for each word. This last part was not being done, resulting in the
possibility of declared i32 register capacities to be off by the number
of 64bit remappings in any CFG.

Fixes the runtime crash in #14313 by generating proper code. However,
additional runtime hardening work is needed:

* The verifier is presently not verifying remapped register ordinals at
all.
* Even if verification is enabled for this, it does not trigger because
it is verifying that they are in range of the maximum allowable (vs
declared).
* There should be padding between the i32/ref register storage areas,
and ASAN marking should be done to catch any overflows.