Remove the ukernels standalone plugin (#14339)
This was done as a stepping stone towards ukernels-as-bitcode. Now that
we have that, we don't need the stepping stone anymore. Like the bitcode
build, this involved custom invocations of custom Clang to produce a
multi-arch library; this is expensive to maintain as it's a whole
separate build of all ukernel code to all target architectures.

The system-plugin remains as it's much simpler (it simply links the
native-build of ukernels as a regular library) and it's much more likely
to still be used by some classes of users going forward (people going to
places where only a system-provided toolchain can go, and not caring for
the multi-arch aspect).