Switch CMake to build libIREECompiler.so without explicit export lists. (#14357)
The need to run generate_exports.py was always a hack that we didn't
want to carry forward. This patch does the CMake plumbing needed to
build the compiler shared library without such a workaround.

It relies on the source level annotations to control what gets exported
in deployment builds (in dev builds, everything is exported on Posix).
This also adds a mechanism for super-projects to extend what gets
exported (which was the motivating use case to fix this).

There is a similar-effect (but different mechanism) way to achieve this
on Bazel, but it requires a Bazel upgrade to use. I'll do that
separately. Until then, the old export lists will be used for Bazel.