[StableHLO] Clean up HLO to Linalg (Part 2/2) (#14462)
This is the first part of the code ported from mlir-hlo which, unlike
latter conversion passes, had not been cleaned up.

Part 1: https://github.com/openxla/iree/pull/14450. This cleanup was
prompted by a latent bug in the casts performed by this code:
https://github.com/openxla/iree/issues/14415.

In this PR:
*  Clean up the remaining code.
* Fix unsafe type conversions / casts. In general, we can't assume that
type converter returns a specific type. This may not be true with
non-default type converters.
*  Remove some uses of `auto` where the type is not immediately clear.
*  Unpack enumerated elements for better naming instead of accessors.
* Make the coding style more uniform with the rest of the stablehlo
input conversion code.