Enable multi-use fusion for cases where the shapes of the two ops are the same. (#13747)
Previous implementation of multi-use fusion was too aggressive making
it hard to enable by default. For example it would fuse a reduction
producer with its consumer which generates outputs of different shapes
making it hard for backends to handle this. Making the conditions for
fusion less aggressive to restrict this to all parallel
producers/consumers of the same shape, should make this handalable on
all backends.

Issue #13545