Adding the `#util.byte_pattern` attribute. (#14556)
This allows for non-splat attributes that are serialized out to files
with predictable values but in the future this should be extented as
documented in #6908 to be backed by a PRNG to approximate model
compression ratios on vmfb files. We could also make the byte pattern
variable width based on the element type but today it's bytes only ala
memset.

Progress on #6908.