[metal] Move to hal/drivers and default build for Apple silicon (#14129)
This commit moves the metal HAL driver from the `experimental/`
directory to the `iree/hal/drivers/`, and turns on building by
default for arm64 Apple devices. x86_64 Apple devices are not
supported right now with missing features and such; that's to
come later.

These existing code is already extensively reviewed and passing
various in-tree tests. Further improvements are for performance
and tracked in https://github.com/openxla/iree/issues/14050.