Making `iree-util-strip-and-splat-constants` use byte_patterns. (#14627)
Had to remove the path someone added for arith.constant as that can't
take byte patterns (and it wasn't the intent of the pass). If someone
wants that they can hoist into globals first before running this pass
with `--iree-util-outline-constants`.

A common usage of this would be to take a giant mlir file and run it
through iree-opt with
`--iree-util-outline-constants --iree-util-strip-and-splat-constants` to
outline any arith.constant values into globals and then strip them.

Note that the final VMFB will contain the same number of bytes as the
original unstripped file.