Delete stale objectives.md document. (#14625)
At one point we made an effort to define quarterly project goals and
publish those on GitHub. We still track work in open source, but we
don't regularly do this level of planning :P

If we had more of these files, I could see keeping them archived by
date. As-is, I'd rather just delete the file and keep it in version
control history only.