Removing initial_data from iree_hal_allocator_allocate_buffer. (#14674)
I didn't really like this when I added it and this reverts it back to
the original behavior of requiring users to allocate and then transfer
whatever they want in their buffers. All implementations besides local
CPU and WebGPU were staging via the extremely inefficient and blocking
synchronous transfer path and we don't want to encourage that - users
must make the decision as to whether they want to synchronously
transfer (in which case it's just an iree_hal_device_transfer_* call)
or asynchronously transfer from files or queue copies.

This likely breaks WebGPU but we can fix that with better solutions -
the existing approach worked a bit but ran down code paths that were
dangerously close to blocking and it didn't account for a lot of buffer
types.

Progress on https://github.com/openxla/iree/issues/14607.
Fixes https://github.com/openxla/iree/issues/14605.