Add service to preload docker images on runner startup (#14196)
This should reduce job latency by pre-fetching some work that would
otherwise be done when we call `docker run`. The disadvantage is that
it will be fetching unnecessary things, so we try to keep the set of
images limited.

This won't reduce resource usage, but should reduce job latency on
average.