ukernels: remove the lone inline-asm ukernel code path (#14853)
Ukernel inner-loop ("tile") funcs used to be written in assembly until
codegen integration underscored the benefits of lifting them to
C+intrinsics unless there was a specific reason not to, and this one was
the one that had such a reason: on ARM Cortex-A510, this inline asm code
achieves 100 Gop/s, compared to only 80 Gop/s for C+intrinsics.

But now, the benefits of going to C+intrinsics are only getting deeper
with ongoing ukernel improvements (narrow matmul kernels mean we are
more interested in parametrizability, and our C->bitcode compilation
path means we get a really nice "templatization" mechanism just by
letting codegen specialize for late-bound constants). And looking closer
at earlier benchmarks... while this asm code was beneficial on the
Cortex-A510, it was slightly detrimental on the Cortex-X2, so it's not
even a pure performance win. And there's a code complexity too -- this
PR drops 350 lines of code, and the inline asm code path was bigger and
more complex than the C+intrinsics code path, not just in that it was
asm, but it was 2x partially unrolled and loop-pipelined.

Note, the ability to use inline-asm in any ukernel remains, and we are
using it for compiler workarounds like here:

https://github.com/openxla/iree/blob/feb548d398c6e4696bfebdfef92da13e413ca5d5/runtime/src/iree/builtins/ukernel/arch/arm_64/mmt4d_arm_64_fp16fml.c#L10-L26
This PR removes infra that we had to let asm and intrinsics variants
coexist side by side with both being tested. That could be revived as
needed.