Upgrade releases to manylinux 2_28. (#14927)
We've been building with this in a downstream for a while. Newer
versions of tools that work better help.

Moving on this now because I hit #14926, which should be fixed, but
reminded me that this old version of manylinux/glibc is missing
performance improving API calls.

Once this lands, we can also switch the deploy script to use clang/lld.