Make byo-bitcode usable wrt multiple targets (#14883)
We've added a `--iree-link-bitcode` flag allowing to BYO bitcode. It
currently allows specifying a single bitcode file that's unconditionally
linked into every executable variant.

That's not usable for a key intended use case, providing
architecture-specific ukernel implementations, as demo'd in
https://github.com/openxla/iree/pull/14496. Concretely the trigger for
this was that @lundong gave that a try and ran into multiply-defined
symbol issues, with the symbol that they were trying to provide a
definition for on RISCV-32. The problem was that because of
constant-evaluation, even though they were targeting RISCV-32, there was
a host executable variant, x86-64, and the custom bitcode file was being
linked there too, clashing with the built-in ukernel bitcode for x86-64.

Because ukernel entry points are actually targeting wasm64/wasm32, so
multiple architectures go through the same wasm architecture for the
entry point, we can't rely on that to skip bitcode modules not relevant
to the current executable variant.

So I don't think there is a way around the need to allow the user to
specify that a bitcode file only applies to a specific architecture.
This PR does that, implementing an optional `arch:path` syntax. The
downside is that the colon (`:`) becomes an illegal character in paths.

Along the way, this PR allows specifying multiple bitcode files, as a
comma-separated list.