[LLVMGPU] Allow reductions with dynamic parallel dimensions through WarpReduce (#14922)
This mirrors what is currently done on the SPIR-V side where we naively
fully distribute dynamic parallel dimensions to workgroups and use the
subgroup reduce path. In the future we will want to use some mix of
padding/specialization for cases like contractions where data is reused.

Additionally, the pipeline matching logic between SPIR-V and LLVMGPU has
nearly converged for subgroup reduction; unifying the two is left as
TODO.

This also disables fused leading elementwise for the contraction
pipelines as this will currently fail to rematerialize the elementwise
and overuse shared memory.