[GlobalOptimization] Fold unit extent dims before fusing dequant with matmul (#15018)
Folding the unit dimensions away for matvec cases in advance eases
codegen lowerings.

Additionally this improves tile size selection for multi-reduce ops with
inputs broadcasted along a faster varying reduction dimension. In such
cases, because we're expecting the broadcasted value to be splatted to
the faster varying reduction dimension we should prefer a vector size of 1.