Add iree-reduce tool (#14892)
This patch adds an `iree-reduce` tool, a test case reduction tool which
reduces the size of a given test case, producing the same error as the
original test case.

The tool implementation is split into 2 parts: Framework and Strategies.
"Framework" is independent of IREE and implements the Delta Debugging
Algorithm.
"Strategies" implements IREE-specific delta passes for test case
reduction.

Currently, 3 delta passes are implemented:
1. ReduceLinalgOnTensorsDelta: Replaces linalg operations with a fill
operation
2. ReduceFlowDispatchOperandToResultDelta: Replace the result of a
flow.dispatch operation with it's operand, if they have the same shape
3. ReduceFlowDispatchResultBySplatDelta: Replace a flow.dispatch
operation with a splat operation.

The delta passes are initial deltas that worked well. They can be
upgraded as we need arises.