[CPU] Improve computation of vector sizes (#14935)
Improve the computation of vector sizes in the vectorizer. We give priority to the vector sizes encoded in the lowering config attributes and use the Value Bounds analysis to retrieve them from the IR if the lowering config information is not provided. Along with this changes, this PR fixes a few bugs that were exposed with the changes and caught by `TilingConfig` verification.