[rocm] Enable the ROCM compiler target backend by default. (#15111)
Includes a small patch to the ROCM bitcode project to disable testing
when included within IREE.

Note: Per the RFC on iree-discuss, we are enabling this by default as an
incremental step in supporting ROCM fully. Enabling the compiler backend
will allow us to more easily get mileage on it in mainline CI and
integration flows, and we may revert if there are problems.