[gpu] Add basic heuristics for better reduction occupancy (#15120)
For subgroup reduction, currently a workgroup processes all elements in
all reduction dimensions. We need to first make sure the workgroup size
we use can divide the total reduction size, and it's also within
hardware limitations.

At the moment we err on the side of using the largest workgroup size
possible; that's good for pure 1-D reduction where we can only ever
launch one workgroup.

Recent work for LLMs we are using subgroup reduction pipeline for matvec
operations--there are parallel dimensions there and can be large. For
such cases, we need to strike a balance--

1) parallel dimensions are distributed to workgroups. If there are
   many workgroups dispatched, we'd want to have each GPU core hosting
   multiple of them for occupancy.
2) we want each thread to read quite a few 128-bit vectors for better
   memory cache behavior.

Both require using a smaller workgroup size. So this commit add basic
heuristics to consider these facts and improve occupancy.